separate token service, update and add tests
