+++ modules/travel_pay/spec/services/token_service_spec.rb
@@ -1,0 +1,105 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe TravelPay::TokenService do
+  let(:user) { build(:user) }
+
+  before do
+    @stubs = Faraday::Adapter::Test::Stubs.new
+
+    conn = Faraday.new do |c|
+      c.adapter(:test, @stubs)
+      c.response :json
+      c.request :json
+    end
+
+    allow_any_instance_of(TravelPay::TokenService).to receive(:connection).and_return(conn)
+  end
+
+  context 'request_veis_token' do
+    it 'returns veis token from proper endpoint' do
+      tenant_id = Settings.travel_pay.veis.tenant_id
+      @stubs.post("#{tenant_id}/oauth2/token") do
+        [
+          200,
+          { 'Content-Type': 'application/json' },
+          '{"access_token": "fake_veis_token"}'
+        ]
+      end
+      token_service = TravelPay::TokenService.new
+      token = token_service.request_veis_token
+
+      expect(token).to eq('fake_veis_token')
+      @stubs.verify_stubbed_calls
+    end
+  end
+
+  context 'request_btsss_token' do
+    before do
+      allow_any_instance_of(TravelPay::TokenService)
+        .to receive(:request_veis_token)
+        .and_return('veis_token')
+      allow_any_instance_of(TravelPay::TokenService)
+        .to receive(:request_sts_token)
+        .and_return('sts_token')
+    end
+
+    it 'returns btsss token from proper endpoint' do
+      @stubs.post('api/v1/Auth/access-token') do
+        [
+          200,
+          { 'Content-Type': 'application/json' },
+          '{"data": {"accessToken": "fake_btsss_token"}}'
+        ]
+      end
+
+      token_service = TravelPay::TokenService.new
+      token = token_service.request_btsss_token(user)
+
+      expect(token).to eq('fake_btsss_token')
+      @stubs.verify_stubbed_calls
+    end
+  end
+
+  context 'request_sts_token' do
+    let(:assertion) do
+      {
+        'iss' => 'https://www.example.com',
+        'sub' => user.email,
+        'aud' => 'https://www.example.com/v0/sign_in/token',
+        'iat' => 1_634_745_556,
+        'exp' => 1_634_745_856,
+        'scopes' => [],
+        'service_account_id' => nil,
+        'jti' => 'c3fa0763-70cb-419a-b3a6-d2563e7b8504',
+        'user_attributes' => { 'icn' => '123498767V234859' }
+      }
+    end
+    let(:grant_type) { 'urn:ietf:params:oauth:grant-type:jwt-bearer' }
+
+    before do
+      Timecop.freeze(Time.zone.parse('2021-10-20T15:59:16Z'))
+      allow(SecureRandom).to receive(:uuid).and_return('c3fa0763-70cb-419a-b3a6-d2563e7b8504')
+    end
+
+    after { Timecop.return }
+
+    it 'builds sts assertion and requests sts token' do
+      private_key_file = Settings.sign_in.sts_client.key_path
+      private_key = OpenSSL::PKey::RSA.new(File.read(private_key_file))
+      jwt = JWT.encode(assertion, private_key, 'RS256')
+      @stubs.post("http:/v0/sign_in/token?assertion=#{jwt}&grant_type=#{grant_type}") do
+        [
+          200,
+          { 'Content-Type': 'application/json' },
+          '{"data": {"access_token": "fake_sts_token"}}'
+        ]
+      end
+      token_service = TravelPay::TokenService.new
+      sts_token = token_service.request_sts_token(user)
+      expect(sts_token).to eq('fake_sts_token')
+      @stubs.verify_stubbed_calls
+    end
+  end
+end
